[Decomp][Spell] Cohesion split: extract SpellEffectSummonLock handlers to SpellEffectSummonLock.cpp
